+++ src/commands/index.ts
@@ -18,3 +18,5 @@ export * from "./inspect";
 export * from "./capture";
 export * from "./impairments";
 export * from "./edgeshark";
+export * from "./openBrowser";
+export * from "./telnet";